feat: add option to modify questions
